+++ src/asp/Core/StereoSettings.h
@@ -292,7 +292,7 @@ namespace asp {
     bool view_matches, view_several_side_by_side, colorize, preview;
     std::string match_file, gcp_file, dem_file, csv_datum, csv_format_str, csv_proj4, nvm;
     bool delete_temporary_files_on_exit;
-    bool create_image_pyramids_only, hide_all;
+    bool create_image_pyramids_only, hide_all, nvm_no_shift;
     bool pairwise_matches, pairwise_clean_matches, no_georef;
     std::vector<std::string> vwip_files;
     vw::BBox2 zoom_proj_win;